feat(notifications): séparer la page en deux — actualités et préférences
- NotificationsPage.vue ne contient plus que le fil d'actualités (filtre
  par manga, liste d'articles, pagination)
- NotificationPreferencesPage.vue reprend le formulaire de préférences
  (sélecteur de canal avec logos, saisie email/webhook, envoi de test)
  ainsi que les alertes système asynchrones
- Nouvelle route /notification-preferences (name: notification-preferences)
- Nouveau lien nav « Préférences » avec l'icône BellCog (lucide)
- i18n : nav.notificationPreferences ajouté FR/EN, nav.notifications
  renommé « Actualités » / « News »
