+++ front/src/i18n/en.json
@@ -4,7 +4,8 @@
     "collection": "Collection",
     "wishlist": "Wishlist",
     "add": "Add",
-    "notifications": "Notifications",
+    "notifications": "News",
+    "notificationPreferences": "Preferences",
     "journal": "Journal",
     "logout": "Logout",
     "settings": "Settings",
