+++ front/src/i18n/fr.json
@@ -4,7 +4,8 @@
     "collection": "Collection",
     "wishlist": "Liste de souhaits",
     "add": "Ajouter",
-    "notifications": "Notifications",
+    "notifications": "Actualités",
+    "notificationPreferences": "Préférences",
     "journal": "Journal",
     "logout": "Déconnexion",
     "settings": "Réglages",
